requirements.txt
bertopic==0.14.1
fitz==0.0.1.dev2
hdbscan==0.8.29
langdetect==1.0.9
nltk==3.8.1
numpy==1.21.5
opencv_python==4.5.5.64
pandas==1.5.2
Pillow==10.0.1
PyMuPDF==1.22.5
pypdf==3.16.2
pypdfium2==4.18.0
pytesseract==0.3.10
requests==2.27.1
scikit_learn==1.3.1
sentence_transformers==2.2.2
tqdm==4.64.0
umap==0.1.1
umap_learn==0.5.3
